Add stream command with worker scaling
This commit adds a new `stream` command to the CLI that allows users to
sip from the firehose. The command includes a `worker-count` flag to
specify the number of workers to scale to. This commit also includes
updates to the `Stream` struct and `BeginStreaming` function to handle
the new command and worker scaling.